+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * [HexaPDF::DigitalSignature::Signing::DefaultHandler#signing_time] for setting
   a custom signing time
 * [HexaPDF::Document#duplicate] for making an in-memory copy of a PDF document
+* Configuration option 'font.default' for setting the default font for the
+  document layout engine
 
 ### Changed